Rename basic operations to algebraic operations.
