+++ symmath/operations/algebraic/conj.hpp
@@ -1,5 +1,5 @@
-#ifndef SYMMATH_OPERATIONS_BASIC_CONJ_HPP
-#define SYMMATH_OPERATIONS_BASIC_CONJ_HPP
+#ifndef SYMMATH_OPERATIONS_ALGEBRAIC_CONJ_HPP
+#define SYMMATH_OPERATIONS_ALGEBRAIC_CONJ_HPP
 
 #include <complex>
 
@@ -58,4 +58,4 @@ inline Conj<T>::Conj(const T &operand)
 
 } // sym
 
-#endif // SYMMATH_OPERATIONS_BASIC_CONJ_HPP
+#endif // SYMMATH_OPERATIONS_ALGEBRAIC_CONJ_HPP
